Added TODO for gcc code generation problem to try and fix
Noticed this staring at the ASM code. This gets used alot, so worth some time trying to find a workaround at some point.

+++ cores/blinklib/blinklib.h
@@ -392,8 +392,9 @@ void loop();
 
 #define FACE_COUNT 6
 
-// 'Cause C ain't got iterators and all those FOR loops are too ugly.
-#define FOREACH_FACE(x) for(int x = 0; x < FACE_COUNT ; ++ x)       // Pretend this is a real language with iterators
+// 'Cause C ain't got no iterators and all those FOR loops are too ugly.
+// TODO: Yuck, gcc expands this loop index to a word, which costs a load, a compare, and a multiply. :/
+#define FOREACH_FACE(x) for(uint8_t x = 0; x < FACE_COUNT ; ++ x)       // Pretend this is a real language with iterators
 
 // Get the number of elements in an array.
 #define COUNT_OF(x) ((sizeof(x)/sizeof(x[0])))
